TST: more complete basic tests (below for details)
- string as num
- basics
- overwriting optional args
- optional args
