+++ tests/optional.yaml
@@ -1,0 +1,12 @@
+name: Test optional arguments
+spec: example.yaml
+load_errors: []
+inputs:
+  basic:
+    function: contact
+    args:
+      name: Scott
+      phone_number: 651
+      email: scott@gmail.com
+    verified:
+      str: Hello
