+++ tests/specs/example.yaml
@@ -6,6 +6,10 @@ contact:
     phone_number:
       type: number
       description: Your phone number, including country code and any area codes
+    email:
+      type: str
+      optional: true
+      default: junk@gmail.com
   rets:
     type: string
     default: Hello
