+++ tests/string_num.yaml
@@ -1,0 +1,11 @@
+name: Test optional arguments
+spec: example.yaml
+load_errors: []
+inputs:
+  basic:
+    function: contact
+    args:
+      name: Scott
+      phone_number: "651"
+    verified:
+      str: Hello
